Resources: Derive `fabric.mod.json` dependency versions from constants
Dependencies in the project are not updated unless something requires
them to change, which usually means that older versions of that
dependency will not work.

Instead of specifying their versions in both `fabric.mod.json` and
`build.gradle.kts`, we should use the `processResources` task to
fill them in at build-time.